[Refactor] test
